fix: use external id in call flow action
