+++ packages/pieces/community/subflows/package.json
@@ -1,4 +1,4 @@
 {
   "name": "@activepieces/piece-subflows",
-  "version": "0.2.1"
+  "version": "0.3.0"
 }
